+++ TestSupport/Assert.Extensions/StringAssertionExtensions.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) 2020 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
+// Licensed under MIT license. See License.txt in the project root for license information.
+
+using System;
 
 namespace Xunit.Extensions.AssertExtensions
 {