+++ TestSupport/EfHelpers/LogOutput.cs
@@ -1,9 +1,7 @@
-﻿// // Copyright (c) 2017 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
-// // Licensed under MIT licence. See License.txt in the project root for license information.
+﻿// Copyright (c) 2020 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
+// Licensed under MIT license. See License.txt in the project root for license information.
 
 using System.Linq;
-using System.Text.RegularExpressions;
-using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.Extensions.Logging;
 using TestSupport.EfHelpers.Internal;
 
@@ -16,20 +14,28 @@ namespace TestSupport.EfHelpers
     {
         private const string EfCoreEventIdStartWith = "Microsoft.EntityFrameworkCore";
 
+        internal LogOutput(LogLevel logLevel, 
+            EventId eventId, string message) 
+        {                                    
+            LogLevel = logLevel;             
+            EventId = eventId;               
+            Message = message;               
+        }
+
         /// <summary>
         /// The logLevel of this log
         /// </summary>
-        public LogLevel LogLevel { get; } 
+        public LogLevel LogLevel { get; }
 
         /// <summary>
         /// The logging EventId - should be string for EF Core logs
         /// </summary>
-        public EventId EventId { get; }  
+        public EventId EventId { get; }
 
         /// <summary>
         /// The message in the log
         /// </summary>
-        public string Message { get; }  
+        public string Message { get; }
 
         /// <summary>
         /// This returns the last part of an EF Core EventId name, or null if the eventId is not an EF Core one
@@ -38,15 +44,7 @@ namespace TestSupport.EfHelpers
             EventId.Name?.StartsWith(                
                      EfCoreEventIdStartWith) == true 
                 ? EventId.Name.Split('.').Last()     
-                : null;                              
-
-        internal LogOutput(LogLevel logLevel, 
-            EventId eventId, string message) 
-        {                                    
-            LogLevel = logLevel;             
-            EventId = eventId;               
-            Message = message;               
-        }                                    
+                : null;
 
         /// <summary>
         /// Summary of the log
@@ -76,6 +74,5 @@ namespace TestSupport.EfHelpers
 
             return EfCoreLogDecoder.DecodeMessage(this);
         }
-
     }
 }