+++ TestSupport/EfHelpers/SqlAdoNetHelpers.cs
@@ -1,5 +1,5 @@
-﻿// Copyright (c) 2017 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
-// Licensed under MIT licence. See License.txt in the project root for license information.
+﻿// Copyright (c) 2020 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
+// Licensed under MIT license. See License.txt in the project root for license information.
 
 using Microsoft.Data.SqlClient;
 
@@ -27,6 +27,7 @@ namespace TestSupport.EfHelpers
                 return (int) myCommand.ExecuteScalar();
             }
         }
+
         /// <summary>
         /// Execute a non-query SQL using ADO.NET
         /// </summary>