+++ TestSupport/EfHelpers/WipeDbViaSql.cs
@@ -1,5 +1,5 @@
-﻿// Copyright (c) 2017 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
-// Licensed under MIT licence. See License.txt in the project root for license information.
+﻿// Copyright (c) 2020 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
+// Licensed under MIT license. See License.txt in the project root for license information.
 
 using System;
 using System.Collections.Generic;
@@ -153,7 +153,5 @@ namespace TestSupport.EfHelpers
                 $"It looked to a depth of {maxDepth} and didn't finish. Possible circular reference?\nentity(s) left: " +
                 string.Join(", ", principalsDictKeys.Select(x => x.ClrType.Name)));
         }
-
-
     }
 }